# docs/api_contract.yaml
openapi: 3.0.3
info:
  title: Tele-Proxy API
  version: 1.0.0
  description: |
    API de proxy de acceso remoto con dos niveles de seguridad:
    - Endpoint público sin autenticación (solo whitelist + rate limit)
    - Endpoints de control con Bearer token
paths:
...
  /api/v1/clients/config:
    get:
      summary: Obtiene la configuración del cliente autenticado
      description: |
        Devuelve la configuración completa del cliente, incluyendo:
        - crypt_id actual (público, rotable)
        - whitelist de dominios
        - rate limit
        - configuración de scripting
        - manejo de errores
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Configuración actual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientConfigResponse"
        "401":
          description: Token Bearer inválido o ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "unauthorized"
                message: "No autorizado: Missing Authorization header"
        "404":
          description: Ningún cliente con ese token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "config_not_found"
                message: "Configuración no encontrada para el cliente: client_123"
        "500":
          description: Fallo de CouchDB u otro error no clasificado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "internal_error"
                message: "Error interno: couchdb view query failed"

    put:
      summary: Actualiza la configuración del cliente
      description: |
        Actualiza la configuración del cliente autenticado.
        Requiere el hash de integridad (`sha256:<64 hex>`) para scripts Lua y fallbacks;
        el `PUT` comprueba el formato, la comparación `sha256(code) == code_hash` ocurre
        en tiempo de ejecución antes de correr el script (ver `spec.md`, Fase 4).
        Incrementa automáticamente el config_version (invalida caché sin purga masiva).
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientConfigUpdate"
      responses:
        "200":
          description: Configuración actualizada exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  config_version:
                    type: integer
                    description: Nueva versión de configuración
                  updated_at:
                    type: string
                    format: date-time
        "400":
          description: >
            Error de validación. `PUT` valida todos los límites del contrato
            (`max_requests 1..=10000`, `window_seconds 1..=3600`,
            `max_scripting_body_bytes 1048576..=52428800`, `whitelist` de hasta 100
            hosts desnudos sin IP privadas, `code_hash` y el `hash` de cada fallback
            en formato `sha256:<64 hex>`) antes de escribir en CouchDB, y también
            devuelve 400 cuando el cuerpo no deserializa (p. ej. un
            `error_handling.mode` fuera de `transparent`/`wrapped`): Axum respondería
            422 en texto plano, y se normaliza a este contrato. Códigos:
            `invalid_config` para las cotas y los hashes, `invalid_url_format` para
            una URL de fallback no http/https.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                out_of_range:
                  summary: Un límite fuera del rango permitido
                  value:
                    error: "invalid_config"
                    message: "Configuración inválida: rate_limit.max_requests 20000 is outside 1..=10000"
                whitelist_entry_not_a_host:
                  summary: Entrada de whitelist que lleva una cabecera o una IP privada
                  value:
                    error: "invalid_config"
                    message: "Configuración inválida: whitelist 'shutterstock.com:8080' is not a bare domain name"
                bad_fallback_url:
                  summary: URL de fallback malformada
                  value:
                    error: "invalid_url_format"
                    message: "Formato de URL inválido: Invalid scheme 'ftp': only http and https are allowed"
        "401":
          description: Token Bearer inválido o ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "unauthorized"
                message: "No autorizado: Invalid Authorization format, expected Bearer"
        "404":
          description: Ningún cliente con ese token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "config_not_found"
                message: "Configuración no encontrada para el cliente: client_123"
        "500":
          description: Fallo de CouchDB u otro error no clasificado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "internal_error"
                message: "Error interno: couchdb write failed"
  /api/v1/clients/rotate-id:
    post:
      summary: Rota el crypt_id público del cliente
      description: |
        Genera un nuevo nanoid de 12 caracteres para el cliente.
        El crypt_id anterior se invalida inmediatamente.
        Útil si se detecta abuso o exposición no deseada del ID público.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Nuevo crypt_id generado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  new_crypt_id:
                    type: string
                    pattern: "^[A-Za-z0-9_-]{12}$"
                    example: "9Kp2mX7qR4wL"
                    description: "Nuevo nanoid de 12 caracteres URL-safe"
                  rotated_at:
                    type: string
                    format: date-time
        "401":
          description: Token Bearer inválido o ausente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "unauthorized"
                message: "No autorizado: Empty bearer token"
        "404":
          description: Ningún cliente con ese token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "config_not_found"
                message: "Configuración no encontrada para el cliente: client_123"
        "500":
          description: Fallo de CouchDB u otro error no clasificado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                error: "internal_error"
                message: "Error interno: couchdb write failed"

  # ========================================
  # ADMINISTRACIÓN GLOBAL (cookie de sesión o master token)
  # ========================================
  # Panel mínimo embebido en GET /admin/ (mismo puerto que la API de control). La autenticación
  # acepta la cookie `teleproxy_admin` (login Google, HttpOnly + SameSite=Lax) o el
  # `MASTER_BEARER_TOKEN` como Bearer — el master SOLO autoriza este router, nunca la API de
  # cliente. Rate limits por IP: login (ADMIN_LOGIN_RATE_LIMIT_*), resto (ADMIN_RATE_LIMIT_*).
  # El código Lua del cliente no aparece en listados; solo en el GET individual.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Token único por cliente para endpoints de control
    AdminSessionCookie:
      type: apiKey
      in: cookie
      name: teleproxy_admin
      description: |
        Sesión de administrador creada por el login Google (HttpOnly, SameSite=Lax, TTL
        `ADMIN_SESSION_TTL_SECONDS`). Alternativa en todas las rutas admin el header
        `Authorization: Bearer <MASTER_BEARER_TOKEN>` (MasterBearer).
    MasterBearer:
      type: http
      scheme: bearer
      description: |
        `MASTER_BEARER_TOKEN` como fallback de emergencia. SOLO autoriza `/api/v1/admin/*`:
        la API de cliente (`/api/v1/clients/*`) y el proxy (`/aq/`) no lo aceptan nunca.

  # Nombres de cabecera en el alambre: Pingora escribe en minúsculas los nombres `&'static str`
  # que se le pasan a `ResponseHeader::insert_header` (pingora-http 0.9.0 conserva la casing
  # recibida), así que las respuestas del proxy llevan `x-cache`, `x-proxy-by`,
  # `x-content-type-options` y `retry-after` en minúsculas. El nombre capitalizado que se usa
  # aquí es el canónico de HTTP/2+OpenAPI: la comparación de cabeceras es case-insensitive.
...
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Código de error del diccionario
          example: "domain_not_whitelisted"
        message:
          type: string
          description: Display del `ProxyError` (dominio/errors.rs); no un texto libre
          example: "Dominio no permitido: malicious.com"
    ClientConfigResponse:
      type: object
      properties:
        crypt_id:
          type: string
          pattern: "^[A-Za-z0-9_-]{12}$"
          example: "V1StGXR8_Z5j"
          description: ID público rotable
        wildcard:
          type: boolean
          default: false
          description: |
            Cliente especial de confianza con salida libre: salta la whitelist de dominios
            y puede salir a CUALQUIER dominio. El anti-SSRF de IP (rechazo de IP privadas,
            pinning DNS, resolución validada) **sigue aplicando igual** — wildcard solo salta
            la capa de whitelist. NO es modificable por el cliente: ni `ClientConfigUpdate`
            ni el `PUT /api/v1/clients/config` lo reciben; solo la API admin
            (`POST/PUT /api/v1/admin/clients`) puede fijarlo o modificarlo.
        config_version:
          type: integer
          description: Versión actual de la configuración
        whitelist:
          type: array
          items:
            type: string
          example: ["shutterstock.com", "gettyimages.com"]
        rate_limit:
          type: object
          properties:
            max_requests:
              type: integer
              minimum: 1
              maximum: 10000
              example: 50
            window_seconds:
              type: integer
              minimum: 1
              maximum: 3600
              example: 60
          description: |
            Mismos límites que `ClientConfigUpdate.rate_limit`: `PUT /api/v1/clients/config`
            los valida y devuelve 400 fuera de rango, así que lo que se lee aquí siempre
            está dentro de ellos.
        max_scripting_body_bytes:
          type: integer
          example: 5242880
          description: Límite de bytes para scripting Lua
        scripting:
          type: object
          properties:
            enabled:
              type: boolean
            code_hash:
              type: string
              description: Hash SHA-256 del script Lua (no se expone el código)
        error_handling:
          type: object
          properties:
            mode:
              type: string
              enum: [transparent, wrapped]
            fallback_urls:
              type: object
              additionalProperties:
                type: object
                properties:
                  url:
                    type: string
                  hash:
                    type: string
        header_rules:
          type: array
          items:
            $ref: "#/components/schemas/HeaderRule"
          description: |
            Reglas de modificación de headers de respuesta. El lenguaje de expresiones y los
            límites completos viven en `docs/HEADER_RULES.md` (sintaxis Cloudflare Rules Engine).
    ClientConfigUpdate:
      type: object
      description: |
        Actualización parcial de la API de cliente (`PUT /api/v1/clients/config`). No incluye
        ni acepta `wildcard` (salida libre) ni `kind`: esos campos son solo-admin y ni siquiera
        deserializan en este payload; activarlos desde el cliente es imposible por construcción.
      type: object
      properties:
        whitelist:
          type: array
          items:
            type: string
          example: ["shutterstock.com", "gettyimages.com"]
        rate_limit:
          type: object
          properties:
            max_requests:
              type: integer
              minimum: 1
              maximum: 10000
            window_seconds:
              type: integer
              minimum: 1
              maximum: 3600
        max_scripting_body_bytes:
          type: integer
          minimum: 1048576
          maximum: 52428800
        scripting:
          type: object
          properties:
            enabled:
              type: boolean
            code:
              type: string
              description: Código Lua completo
            code_hash:
              type: string
              description: Hash SHA-256 del código (obligatorio si se proporciona code)
        error_handling:
          type: object
          properties:
            mode:
              type: string
              enum: [transparent, wrapped]
            fallback_urls:
              type: object
              additionalProperties:
                type: object
                required:
                  - url
                  - hash
                properties:
                  url:
                    type: string
                    format: uri
                  hash:
                    type: string
                    description: Hash SHA-256 del recurso fallback (obligatorio)
        header_rules:
          type: array
          items:
            $ref: "#/components/schemas/HeaderRule"
          description: |
            Sustituye la lista completa de reglas (como `whitelist`); `[]` las desactiva.
            Máximo 50 reglas. Cada expresión se valida (sintaxis, tipos, regex de `matches`)
            antes de escribir: un error devuelve 400 `invalid_config` con el campo concreto,
            p. ej. `header_rules[0].expression`.
...